add-viewlib: rename templateEditor-mgr to template-editor-mgr
