+++ app/view/netcreate/template-editor-mgr.js
